hot reload for all lua handlers
Hot reload code is now integrated into cfg(), this way there is less
copy-paste and we can check everything for sanity (not everything is
checked yet for easier merging). All Lua handlers are now replaced but
this is not being done atomically or efficiently.

Some parameters from new configuration are used before everything is
checked for sanity (to be improved later).

Also fixed some memory/resource leaks when initialization fails due to
invalid parameters.

Part of #18